test(arm64): widen arch-neutral wago tests to run on arm64 (ledger AMD64→ARM64 #7)
The linux&&amd64 wago suite shares pure public-API exec helpers (runv/run1/
runImports) that only its build tag kept off arm64. Move them to a broadly-tagged
exec_shared_test.go, then widen global_test.go (40 tests) and
defer_bounds_checks_test.go to run on arm64 too — both pass on darwin/arm64 and
still on linux/amd64. Remaining files need a small per-file helper untangling
(memprogWasm/testdata, tableTest*/passiveDataModule dedup vs the curated
darwin_arm64_test_helpers_test.go) before they can widen.
