+++ src/wago/defer_bounds_checks_test.go
@@ -1,4 +1,4 @@
-//go:build linux && amd64
+//go:build ((linux && amd64) || arm64) && !tinygo
 
 package wago
